+++ server/src/controllers/admin/adminLogin.controller.js
@@ -1,4 +1,4 @@
-import { Admin } from "../../models/admin.model.js";
+import { User } from "../../models/user.model.js";
 import ApiError from "../../utils/ApiError.js";
 import ApiResponse from "../../utils/ApiResponse.js";
 import asyncHandler from "../../utils/asyncHandler.js";
@@ -10,9 +10,9 @@ export const adminLogin = asyncHandler(async (req, res) => {
     if (!email) throw new ApiError(400, false, "Email not provided");
     if (!password) throw new ApiError(400, false, "Password not provided");
 
-    const admin = await Admin.findOne({ email });
+    const user = await User.findOne({ email });
 
-    if (!admin)
+    if (!user)
         throw new ApiError(
             400,
             false,
