+++ server/src/controllers/index.js
@@ -6,7 +6,7 @@ import { verifyToken } from "./verifyToken.controller.js";
 import { currentUserInfo } from "./currentUserInfo.controller.js";
 import { uploadKyc } from "./uploadKyc.controller.js";
 import { adminLogin } from "./admin/adminLogin.controller.js";
-import { createAdmin } from "./admin/createAdmin.controller.js";
+import { addAdmin } from "./admin/addAdmin.controller.js";
 
 export {
     login,
@@ -18,5 +18,5 @@ export {
     currentUserInfo,
     uploadKyc,
     adminLogin,
-    createAdmin,
+    addAdmin,
 };
